hotel_lister_backend/requirements.txt
Django>=5.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
mysqlclient>=2.1.0
requests>=2.28.0
amadeus>=8.0.0